Revert "[CT-708] Indexer track e2e latency (#1237)"
This reverts commit 60b94dfd6599f367f02f7521cad0e5dbf761d8b7.
