+++ indexer/services/ender/__tests__/helpers/indexer-proto-helpers.ts
@@ -50,7 +50,6 @@ import {
   IndexerOrderId,
   PerpetualMarketCreateEventV1,
   DeleveragingEventV1,
-  protoTimestampToDate,
 } from '@dydxprotocol-indexer/v4-protos';
 import {
   PerpetualMarketType,
@@ -64,6 +63,7 @@ import {
   generatePerpetualMarketMessage,
   generatePerpetualPositionsContents,
 } from '../../src/helpers/kafka-helper';
+import { protoTimestampToDate } from '../../src/lib/helper';
 import { DydxIndexerSubtypes, VulcanMessage } from '../../src/lib/types';
 
 // TX Hash is SHA256, so is of length 64 hexadecimal without the '0x'.
