+++ indexer/services/socks/src/lib/message-forwarder.ts
@@ -3,7 +3,6 @@ import {
   logger,
   InfoObject,
   safeJsonStringify,
-  STATS_NO_SAMPLING,
 } from '@dydxprotocol-indexer/base';
 import { updateOnMessageFunction } from '@dydxprotocol-indexer/kafka';
 import { KafkaMessage } from 'kafkajs';
@@ -86,10 +85,9 @@ export class MessageForwarder {
   }
 
   public onMessage(topic: string, message: KafkaMessage): void {
-    const start: number = Date.now();
     stats.timing(
       `${config.SERVICE_NAME}.message_time_in_queue`,
-      start - Number(message.timestamp),
+      Date.now() - Number(message.timestamp),
       config.MESSAGE_FORWARDER_STATSD_SAMPLE_RATE,
       {
         topic,
@@ -127,31 +125,18 @@ export class MessageForwarder {
       return;
     }
 
-    const startForwardMessage: number = Date.now();
+    const start: number = Date.now();
     this.forwardMessage(messageToForward);
     const end: number = Date.now();
     stats.timing(
       `${config.SERVICE_NAME}.forward_message`,
-      end - startForwardMessage,
+      end - start,
       config.MESSAGE_FORWARDER_STATSD_SAMPLE_RATE,
       {
         topic,
         channel: String(channel),
       },
     );
-
-    const originalMessageTimestamp = message.headers?.message_received_timestamp;
-    if (originalMessageTimestamp !== undefined) {
-      stats.timing(
-        `${config.SERVICE_NAME}.message_time_since_received`,
-        startForwardMessage - Number(originalMessageTimestamp),
-        STATS_NO_SAMPLING,
-        {
-          topic,
-          event_type: String(message.headers?.event_type),
-        },
-      );
-    }
   }
 
   public forwardMessage(message: MessageToForward): void {
