+++ indexer/services/vulcan/__tests__/handlers/order-remove-handler.test.ts
@@ -5,9 +5,6 @@ import {
   STATS_FUNCTION_NAME,
   wrapBackgroundTask,
 } from '@dydxprotocol-indexer/base';
-import {
-  defaultTime,
-} from '../helpers/constants';
 import { synchronizeWrapBackgroundTask } from '@dydxprotocol-indexer/dev';
 import {
   ORDERBOOKS_WEBSOCKET_MESSAGE_VERSION,
@@ -55,16 +52,14 @@ import {
   OrderRemoveV1_OrderRemovalStatus,
   RedisOrder,
   SubaccountMessage,
-  protoTimestampToDate,
 } from '@dydxprotocol-indexer/v4-protos';
 import Big from 'big.js';
-import { IHeaders, ProducerRecord } from 'kafkajs';
+import { ProducerRecord } from 'kafkajs';
 import { DateTime } from 'luxon';
 
 import { OrderRemoveHandler } from '../../src/handlers/order-remove-handler';
 import { OrderbookSide } from '../../src/lib/types';
 import { redisClient } from '../../src/helpers/redis/redis-controller';
-
 import {
   expectCanceledOrderStatus,
   expectOpenOrderIds,
@@ -138,10 +133,6 @@ describe('OrderRemoveHandler', () => {
     timeInForce: TimeInForce.IOC,
   };
 
-  const defaultKafkaHeaders: IHeaders = {
-    message_received_timestamp: String(protoTimestampToDate(defaultTime)),
-  };
-
   it.each([
     [
       {
@@ -188,10 +179,7 @@ describe('OrderRemoveHandler', () => {
     const offChainUpdate: OffChainUpdateV1 = orderRemoveToOffChainUpdate(orderRemoveJson);
 
     const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-    await expect(orderRemoveHandler.handleUpdate(
-      offChainUpdate,
-      defaultKafkaHeaders,
-    )).rejects.toThrow(
+    await expect(orderRemoveHandler.handleUpdate(offChainUpdate)).rejects.toThrow(
       new ParseMessageError(errorMessage),
     );
     expect(logger.error).toHaveBeenCalledWith(expect.objectContaining({
@@ -206,10 +194,7 @@ describe('OrderRemoveHandler', () => {
       const offChainUpdate: OffChainUpdateV1 = orderRemoveToOffChainUpdate(defaultOrderRemove);
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       expect(logger.info).toHaveBeenCalledWith(expect.objectContaining({
         at: 'orderRemoveHandler#handleOrderRemoval',
@@ -230,10 +215,7 @@ describe('OrderRemoveHandler', () => {
 
       const offChainUpdate: OffChainUpdateV1 = orderRemoveToOffChainUpdate(defaultOrderRemove);
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       const ticker: string = testConstants.defaultPerpetualMarket.ticker;
       expect(logger.error).toHaveBeenCalledWith({
@@ -338,10 +320,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       // orderbook level reduced by defaultQuantums
       const remainingOrderbookLevel: string = Big(
@@ -491,10 +470,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       await Promise.all([
         expectOrderStatus(expectedOrderUuid, OrderStatus.BEST_EFFORT_CANCELED),
@@ -628,10 +604,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         await Promise.all([
           expectOrderStatus(expectedOrderUuid, OrderStatus.CANCELED),
@@ -769,10 +742,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         await Promise.all([
           expectOrderStatus(expectedOrderUuid, OrderStatus.CANCELED),
@@ -925,10 +895,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         await Promise.all([
           expectOrderStatus(expectedOrderUuid, removedOrder.status),
@@ -1011,10 +978,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         await Promise.all([
           expectOrderStatus(expectedOrderUuid, OrderStatus.FILLED),
@@ -1050,10 +1014,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       expect(producerSendSpy).not.toHaveBeenCalled();
       expect(logger.error).toHaveBeenCalledWith(expect.objectContaining({
@@ -1101,10 +1062,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       // Subaccounts message is sent first followed by orderbooks message
       const subaccountContents: SubaccountMessageContents = {
@@ -1204,10 +1162,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       await Promise.all([
         // orderbook should not be affected, so it will be set to defaultQuantums
@@ -1324,10 +1279,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       // orderbook level reduced by defaultQuantums
       const remainingOrderbookLevel: string = Big(
@@ -1471,10 +1423,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       await Promise.all([
         expectOrderbookLevelCache(
@@ -1596,10 +1545,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       // orderbook level reduced by defaultQuantums
       const remainingOrderbookLevel: string = Big(
@@ -1730,10 +1676,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       // orderbook level should not be reduced
       const remainingOrderbookLevel: string = Big(
@@ -1795,10 +1738,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         expect(producerSendSpy).not.toHaveBeenCalled();
         expect(logger.error).toHaveBeenCalledWith(expect.objectContaining({
@@ -1865,10 +1805,7 @@ describe('OrderRemoveHandler', () => {
         const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
         const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-        await orderRemoveHandler.handleUpdate(
-          offChainUpdate,
-          defaultKafkaHeaders,
-        );
+        await orderRemoveHandler.handleUpdate(offChainUpdate);
 
         expect(producerSendSpy).not.toHaveBeenCalled();
         expect(stats.increment).toHaveBeenCalledWith('vulcan.indexer_expired_order_not_found', 1);
@@ -1941,10 +1878,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       expect(producerSendSpy).not.toHaveBeenCalled();
       expect(logger.error).toHaveBeenCalledWith(expect.objectContaining({
@@ -2006,10 +1940,7 @@ describe('OrderRemoveHandler', () => {
       const producerSendSpy: jest.SpyInstance = jest.spyOn(producer, 'send').mockReturnThis();
 
       const orderRemoveHandler: OrderRemoveHandler = new OrderRemoveHandler();
-      await orderRemoveHandler.handleUpdate(
-        offChainUpdate,
-        defaultKafkaHeaders,
-      );
+      await orderRemoveHandler.handleUpdate(offChainUpdate);
 
       expect(producerSendSpy).not.toHaveBeenCalled();
       expect(
@@ -2113,11 +2044,7 @@ describe('OrderRemoveHandler', () => {
 
     if (expectedSubaccountMessage !== undefined) {
       const subaccountProducerRecord: ProducerRecord = producerSendSpy.mock.calls[0][0];
-      expectWebsocketSubaccountMessage(
-        subaccountProducerRecord,
-        expectedSubaccountMessage,
-        defaultKafkaHeaders,
-      );
+      expectWebsocketSubaccountMessage(subaccountProducerRecord, expectedSubaccountMessage);
     }
 
     if (expectedOrderbookMessage !== undefined) {
