+++ indexer/services/vulcan/src/handlers/order-place-handler.ts
@@ -32,7 +32,7 @@ import {
   RedisOrder,
 } from '@dydxprotocol-indexer/v4-protos';
 import Big from 'big.js';
-import { IHeaders, Message } from 'kafkajs';
+import { Message } from 'kafkajs';
 
 import config from '../config';
 import { redisClient } from '../helpers/redis/redis-controller';
@@ -52,7 +52,7 @@ import { Handler } from './handler';
  *   being greater than or equal to the expiry of the order in the OrderPlace message, return
  */
 export class OrderPlaceHandler extends Handler {
-  protected async handle(update: OffChainUpdateV1, headers: IHeaders): Promise<void> {
+  protected async handle(update: OffChainUpdateV1): Promise<void> {
     logger.info({
       at: 'OrderPlaceHandler#handle',
       message: 'Received OffChainUpdate with OrderPlace.',
@@ -142,7 +142,7 @@ export class OrderPlaceHandler extends Handler {
           });
           throw new Error(`Stateful order not found in database: ${orderUuid}`);
         }
-        await this.sendCachedOrderUpdate(orderUuid, headers);
+        await this.sendCachedOrderUpdate(orderUuid);
       }
       const subaccountMessage: Message = {
         value: createSubaccountWebsocketMessage(
@@ -151,9 +151,6 @@ export class OrderPlaceHandler extends Handler {
           perpetualMarket,
           placementStatus,
         ),
-        headers: {
-          message_received_timestamp: headers.message_received_timestamp,
-        },
       };
       sendMessageWrapper(subaccountMessage, KafkaTopics.TO_WEBSOCKETS_SUBACCOUNTS);
     }
@@ -167,9 +164,6 @@ export class OrderPlaceHandler extends Handler {
           perpetualMarket,
           updatedQuantums,
         ),
-        headers: {
-          message_received_timestamp: headers.message_received_timestamp,
-        },
       };
       sendMessageWrapper(orderbookMessage, KafkaTopics.TO_WEBSOCKETS_ORDERBOOKS);
     }
@@ -340,7 +334,6 @@ export class OrderPlaceHandler extends Handler {
    */
   protected async sendCachedOrderUpdate(
     orderId: string,
-    headers: IHeaders,
   ): Promise<void> {
     const cachedOrderUpdate: OrderUpdateV1 | undefined = await StatefulOrderUpdatesCache
       .removeStatefulOrderUpdate(
@@ -358,10 +351,6 @@ export class OrderPlaceHandler extends Handler {
       value: Buffer.from(
         Uint8Array.from(OffChainUpdateV1.encode({ orderUpdate: cachedOrderUpdate }).finish()),
       ),
-      headers: {
-        message_received_timestamp: headers.message_received_timestamp,
-        event_type: String(headers.event_type),
-      },
     };
     sendMessageWrapper(orderUpdateMessage, KafkaTopics.TO_VULCAN);
   }
