+++ indexer/services/vulcan/src/lib/on-message.ts
@@ -62,19 +62,6 @@ export async function onMessage(message: KafkaMessage): Promise<void> {
     },
   );
 
-  const originalMessageTimestamp = message.headers?.message_received_timestamp;
-  if (originalMessageTimestamp !== undefined) {
-    stats.timing(
-      `${config.SERVICE_NAME}.message_time_since_received`,
-      start - Number(originalMessageTimestamp),
-      STATS_NO_SAMPLING,
-      {
-        topic: KafkaTopics.TO_VULCAN,
-        event_type: String(message.headers?.event_type),
-      },
-    );
-  }
-
   const messageValue: Buffer = message.value;
   const offset: string = message.offset;
   let update: OffChainUpdateV1;
@@ -99,20 +86,7 @@ export async function onMessage(message: KafkaMessage): Promise<void> {
     const handler: Handler = new (getHandler(update))!(
       getTransactionHashFromHeaders(message.headers),
     );
-    await handler.handleUpdate(update, message.headers ?? {});
-
-    const postProcessingTime: number = Date.now();
-    if (originalMessageTimestamp !== undefined) {
-      stats.timing(
-        `${config.SERVICE_NAME}.message_time_since_received_post_processing`,
-        postProcessingTime - Number(originalMessageTimestamp),
-        STATS_NO_SAMPLING,
-        {
-          topic: KafkaTopics.TO_VULCAN,
-          event_type: String(message.headers?.event_type),
-        },
-      );
-    }
+    await handler.handleUpdate(update);
 
     success = true;
   } catch (error) {
